New branch: split-into-files
This branch aims to refactor the structure of the 'checkocr.py' module.

The main direction should be to create more reusable functions, stored
in one or more files, as a way to simplify the main content-related
methods of the current 'File' class.

With the design of a full suite of unit tests in mind, the code has
to be made clearer with smaller and cleaner methods and functions.

This is also meant to be better equiped to deal with some of the
subtle bottlenecks of the pdf file filtering "algorithm" and
make this filtering process and the role of each of its steps
more precise.